% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{ggmapaRD}
\alias{ggmapaRD}
\title{Graficar mapa de la Rep. Dom.}
\usage{
ggmapaRD(nivel = "provincial", df, var, sub, idName)
}
\arguments{
\item{nivel}{nivel territorial/administrativo (\code{"regional"}, \code{"provincial"}, o \code{"municipal"})}

\item{df}{dataframe con los valores}

\item{var}{nombre de variable}

\item{sub}{subconjunto de territorios}
}
\value{
Mapa formato objeto ggplot
}
\description{
Grafica el mapa de la Rep. Dom. por nivel administrativo (regional,
provincial y municipal).
}
\examples{
d <- data.frame(ID = 1:32, x = rnorm(32))
ggmapaRD("provincial", df = d, var= "x", idName = "ID2")

}
