% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{mapaRD_interactivo}
\alias{mapaRD_interactivo}
\title{Graficar mapa interactivo de la Rep. Dom.}
\usage{
mapaRD_interactivo(df, var, nivel = "provincial", na.rm = FALSE, idName)
}
\arguments{
\item{df}{dataframe con los valores}

\item{var}{nombre de variable}

\item{nivel}{nivel territorial/administrativo (\code{"regional"}, \code{"provincial"}, o \code{"municipal"})}

\item{na.rm}{logical que cuando es verdadero, excluye del gráfico los territorios
con valores NA}
}
\value{
Mapa formato objeto plotly
}
\description{
Grafica un mapa interactivo de la Rep. Dom. por nivel administrativo (regional,
provincial y municipal).
}
\examples{
d <- data.frame(ID = 1:32, x = rnorm(32))
mapaRD_interactivo("provincial", df = d, var= "x", idName = "ID2")


}
