% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{buscarMapaRD}
\alias{buscarMapaRD}
\title{Graficar mapa de la Rep. Dom.}
\usage{
buscarMapaRD(nivel, sub)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin parameters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Grafica el mapa de la Rep. Dom. por distintos niveles administrativos: regional,
provincial y municipal.
}
