% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{buscarMapaRD}
\alias{buscarMapaRD}
\title{Buscar coordenadas de mapas por division territorial}
\usage{
buscarMapaRD(nivel, sub, maptype = "ggplot")
}
\arguments{
\item{nivel}{nivel territorial/administrativo (\code{"regional"}, \code{"provincial"}, o \code{"municipal"})}

\item{sub}{subconjunto de territorios}

\item{maptype}{busca el archivo dependiento el tipo de grafico}
}
\value{
None
}
\description{
Función interna al paquete que busca en la data las coordenadas de los mapas
por territorio
}
\examples{
buscarMapaRD(nivel="provincial")

}
