% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\docType{data}
\name{tabla_toponimia}
\alias{tabla_toponimia}
\title{Tabla de toponimia}
\format{
An object of class \code{list} of length 3.
}
\usage{
tabla_toponimia
}
\arguments{
\item{nivel}{nivel territorial/administrativo (\code{"regional"}, \code{"provincial"}, o \code{"municipal"})}
}
\value{
data.frame
}
\description{
Muestra la tabla de nombres, identificadores, y demás detalles de
las divisiones territoriales para cada nivel administrativo.
}
\examples{
tabla_toponimia(nivel="provincial")

}
\keyword{datasets}
